+++ src/main/java/raziel23x/projectskyblock/machine/thermal/IThermalFuelConsumer.java
@@ -1,0 +1,5 @@
+package raziel23x.projectskyblock.machine.thermal;
+
+public interface IThermalFuelConsumer {
+    int consumeThermalFuel(int maxAmount, boolean simulate);
+}