+++ src/main/java/raziel23x/projectskyblock/machine/thermal/IThermalFuelProducer.java
@@ -1,0 +1,7 @@
+package raziel23x.projectskyblock.machine.thermal;
+
+public interface IThermalFuelProducer {
+    int getThermalFuelStored();
+
+    int getThermalFuelCapacity();
+}